requirements.txt
playwright==1.41.0
beautifulsoup4==4.12.3
tinydb==4.8.0
spacy==3.7.2
spacytextblob==4.0.0
textblob==0.18.0
scikit-learn==1.4.0
pyyaml==6.0.1
flask==3.0.0
python-dotenv==1.0.0
pandas==2.2.0

